ai/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Maciej Kaszkowiak",email = "maciej.kaszkowiak@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.10,<0.116.0)",
    "uuid7 (>=0.1.0,<0.2.0)",
    "openai (>=1.65.4,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "google-genai (>=1.14.0,<2.0.0)",
    "anthropic (>=0.51.0,<0.52.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
